+++ Clove/src/Graphics/DirectX-11/DX11RenderAPI.hpp
@@ -53,7 +53,7 @@ namespace clv::gfx{
 		virtual void setRenderTarget(RenderTarget& renderTarget) override;
 		virtual void resetRenderTarget() override;
 
-		virtual void setViewportSize(float width, float height) override;
+		virtual void setViewportSize(uint32 width, uint32 height) override;
 
 		static ID3D11Device& getDevice();
 		static ID3D11DeviceContext& getContext();
